requirements.dev.txt
djangorestframework==3.14.0
python-dotenv==1.0.0
drf_yasg==1.21.6
faker
djongo==1.3.7
mongo==0.2.0